Fixed AuthenticationRestControllerIT failing test
